+++ lib/api_validation/VerifiablePresentationTokenValidator.ts
@@ -24,7 +24,7 @@ export default class VerifiablePresentationTokenValidator implements ITokenValid
    * @param validatorOption The options used during validation
    * @param expected values to find in the token to validate
    */
-   constructor (private validatorOption: IValidatorOptions, private expected: IExpectedVerifiablePresentation ) {
+  constructor(private validatorOption: IValidatorOptions, private expected: IExpectedVerifiablePresentation) {
   }
 
   /**
@@ -33,7 +33,7 @@ export default class VerifiablePresentationTokenValidator implements ITokenValid
    * @param queueItem under validation
    * @param siopDid needs to be equal to audience of VP
    */
-  public async validate(queue: ValidationQueue, queueItem: ValidationQueueItem, siopDid: string): Promise<IValidationResponse> { 
+  public async validate(queue: ValidationQueue, queueItem: ValidationQueueItem, siopDid: string): Promise<IValidationResponse> {
     const options = new ValidationOptions(this.validatorOption, TokenType.verifiablePresentationJwt);
     const validator = new VerifiablePresentationValidation(options, this.expected, siopDid, queueItem.id);
     let validationResult = await validator.validate(<string>queueItem.tokenToValidate.rawToken);
@@ -44,13 +44,13 @@ export default class VerifiablePresentationTokenValidator implements ITokenValid
 
     return validationResult as IValidationResponse;
   }
-  
+
   /**
    * Get tokens from current item and add them to the queue.
    * @param validationResponse The response for the requestor
    * @param queue with tokens to validate
    */
-  public getTokens(validationResponse: IValidationResponse, queue: ValidationQueue ): IValidationResponse {
+  public getTokens(validationResponse: IValidationResponse, queue: ValidationQueue): IValidationResponse {
     if (!validationResponse.payloadObject.vp || !validationResponse.payloadObject.vp.verifiableCredential) {
       return {
         result: false,
@@ -60,13 +60,37 @@ export default class VerifiablePresentationTokenValidator implements ITokenValid
       };
     }
 
-    const vc = validationResponse.payloadObject.vp.verifiableCredential;
+    // Validation gate: Test for maximum number of VCs and maximum size
+    const vc: any[] = validationResponse.payloadObject.vp.verifiableCredential;
     validationResponse.tokensToValidate = {};
-    for (let token in vc) {
-      const claimToken = ClaimToken.create(vc[token]);
-      const vcType = VerifiableCredentialValidation.getVerifiableCredentialType(claimToken.decodedToken.vc || claimToken.decodedToken);
-      validationResponse.tokensToValidate[vcType] = claimToken; 
-      queue.enqueueToken(vcType, claimToken);    
+    if (vc) {
+      // Validation gate: Test for maximum number of VPs and maximum size
+      const validatorSafeguards = this.validatorOption.validationSafeguards;
+      if (vc.length > validatorSafeguards.maxNumberOfVCTokensInPresentation) {
+        return {
+          result: false,
+          status: 403,
+          code: errorCode(2),
+          detailedError: `The number of VCs ${vc.length} in presentation exceeds the maximum ${validatorSafeguards.maxNumberOfVCTokensInPresentation}.`
+        }
+      }
+
+      let oversized = vc.filter((token: string) => token.length > validatorSafeguards.maxSizeOfVCTokensInPresentation);
+      if (oversized.length !== 0) {
+        return {
+          result: false,
+          status: 403,
+          code: errorCode(3),
+          detailedError: `The presentation has an oversized VC tokens. Size ${oversized[0].length}. Maximum size: ${validatorSafeguards.maxSizeOfVCTokensInPresentation}.`
+        }
+      }
+
+      for (let token in vc) {
+        const claimToken = ClaimToken.create(vc[token]);
+        const vcType = VerifiableCredentialValidation.getVerifiableCredentialType(claimToken.decodedToken.vc || claimToken.decodedToken);
+        validationResponse.tokensToValidate[vcType] = claimToken;
+        queue.enqueueToken(vcType, claimToken);
+      }
     }
 
     return validationResponse;