+++ lib/options/BasicValidatorOptions.ts
@@ -2,7 +2,7 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { Crypto, IDidResolver, CryptoBuilder, IFetchRequest, FetchRequest } from '../index';
+import { Crypto, IDidResolver, CryptoBuilder, IFetchRequest, FetchRequest, ValidationSafeguards } from '../index';
 import IValidatorOptions from './IValidatorOptions';
 
 /**
@@ -12,19 +12,28 @@ export default class BasicValidatorOptions implements IValidatorOptions {
 
   private readonly _crypto: Crypto;
   private readonly _fetchRequest: IFetchRequest;
+  private readonly _validationSafeguards: ValidationSafeguards;
 
   constructor(private _resolver?: IDidResolver) {
     this._crypto = new CryptoBuilder().build();
     this._fetchRequest = new FetchRequest();
+    this._validationSafeguards = new ValidationSafeguards();
   }
 
   get resolver(): IDidResolver {
     return this._resolver!;
   }
 
-    /**
-     * The fetch client
-     */
+  /**
+   * The fetch client
+   */
+  get validationSafeguards(): ValidationSafeguards {
+    return this._validationSafeguards;
+  }
+
+  /**
+   * The fetch client
+   */
   get fetchRequest(): IFetchRequest {
     return this._fetchRequest;
   }